requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.12
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
whitenoise==6.0.0
zope.interface==5.4.0
